requirements.txt
# Core
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0

# MiniGrid
minigrid>=2.3.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Logging and visualization
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0

# Utilities
python-dotenv>=1.0.0